For back/add

See merge request s09-fintech-finance-sub2/S09P22A409!45

+++ back/seesaw/src/main/java/com/doacha/seesaw/model/dto/MonthCategoryResponse.java
@@ -11,5 +11,5 @@ public class MonthCategoryResponse {
     private long spendingCostSum;
     private int spendingMonth;
     private String memberEmail;
-    private long spendingCategoryId;
+    private int spendingCategoryId;
 }
